requirements.txt
asgiref==3.4.1
boto3==1.18.43
botocore==1.21.43
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
click==8.0.1
colorama==0.4.4
config==0.5.1
cryptography==3.4.8
decorator==5.1.0
Django==3.2.7
django-crispy-forms==1.12.0
django-rest-authtoken==2.1.4
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
dynaconf==3.1.7
hvac==0.11.0
idna==3.2
Jinja2==3.0.1
jmespath==0.10.0
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
Pillow==8.3.2
psycopg2==2.9.1
pycparser==2.20
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
validators==0.18.2
win32-setctime==1.0.3
